Create TradeStrategy
This module will be used to handle the basic buy and sell orders. A neural network will be used to predict the likely prices and adjust the buy/sell orders.  I am hoping this will increase the frequency of trades and thus giving the bot more opportunities to make a profit.